requirements.txt
elasticsearch<8.0.0
google-cloud-storage==1.43.0
openpyxl==3.0.9
pandas==1.4.1
python-dotenv==0.19.2
